Prefer clang over gcc for CC/CXX
The Android NDK uses clang by default, but offers both clang and a
gcc binary; the gcc does not use the correct standard library so
programs will fail to compile or link if it is used instead of
clang. Generally, when clang is provided there is a gcc backup,
but not the other way around, so prefer clang in all cases.