Fix worker crashes when UTF-8 is printed to console
This happens with GCW0's packager, which replaces UTF-8 characters
so prints them to the console, but is a problem generally because
the Docker containers have no set locale that Buildbot can use.
Until the next version of Buildbot (0.9.13), which shouldn't crash
when it receives unexpected data encodings, it seems advisable to
simply have all the workers use UTF-8 by default since this is the
default universal encoding nowadays.