Fixes to cross-compiler compilation
The Vita cross-compiler has third-party dependencies with Makefiles
which assume all port libraries are installed into the host triple
subdirectory. Other ports also seem to be designed this way, like
the Mingw-w64 port on Debian which installs to /usr/<host triple>,
and previously were using some unpleasant hacks around PATH and
deleting the unprefixed binaries in the <host triple>/bin directory
to avoid confusing the host's native compiler with the
cross-compiler. This changeset fixes up compile-libraries.sh to
find the cross-compiler binaries from $PATH, and then fixes all of
the knock-on effects of this change in the various other ports
which had previously been compiling libraries directly into their
sysroots.

This changeset also changes all the workers to base their
compilation layers on the base OS image instead of the common
worker base image, since (obviously, in hindsight) it should not be
necessary to recompile all of the cross-compilers from scratch
every time the Buildbot worker needs to be updated.