Always run configuration once after builder is regenerated
If the buildbot.cfg was changed for a builder, `configure` needs to
run again even if nothing in the source tree changed. It's simple
enough to do this just by marking whether the configuration step
has run once for the builder, since this flag will get reset
whenever the buildmaster is restarted or reconfigured.