Improve library build scripts
This removes most of the hacks for building libraries and brings
support for some libraries that had trouble building before on
some platforms (like flac on Wii) due to those hacks not working
very well. It also cleans up some of the port-specific library
rules so their extra features are more clearly demarcated.