Split Apple worker into separate iOS and macOS workers
While the two are similar, getting the SDK for iOS and setting up
the libraries is so different than it is with macOS that they are
really two separate workers, so now are set up as such.

The macOS worker is now fixed and builds binaries which run; the
iOS worker needs help because the previously used SDK download
location has disappeared from the internet.