Remove SamsungTV worker
This builds, but the glibc used for linking is too new and does not
support the ancient 2.6-era kernel used by these devices, so it
will not actually run on the hardware.

I would like to figure out how to build a modern glibc with support
for kernel 2.6 since this will be needed if Maemo remains a
supported platform, but this will require more work than can be
addressed at this moment.