Don't fail packaging if an unstrippable binary is received
This is needed for PSP which sends an EBOOT.PBP for packaging in
place of an actual binary. All these extra mechanics feel kind of
hacky, so maybe the way PSP does stuff is wrong and needs to be
changed, and in the interest of getting Buildbot wrapped up and
deployed so I can move on to other things, this is what is
happening for now.