Remove most ‘production’ code
Production deployment will be handled separately using Ansible,
since the host machine needs to be configured as well.

For the moment, docker-compose.prod.yml is not removed entirely
because Ansible’s docker_service does not currently let us set
restart policy, and it’s easier to keep track of the restart policy
in the same place as the rest of the Docker configuration. This
will hopefully go away in the future as well, either because the
policy can be set through Ansible, or because the production
overrides just end up being managed by Ansible.