Rename internal Buildbot builders network
When using a transparent tunnel service for workers to communicate
with a buildmaster on another host, it is necessary to replace the
master service without inheriting any of the things from the actual
service configuration, but it is not possible with docker-compose
to just totally replace a service, so we need to hack around this
in Ansible by giving it some unique string to replace when
deploying the tunnel service. Since buildbot is also the name of
the internal network, it was not possible to easily disable the
service by just changing its name, so now the network is named
differently than the service so the service can be targeted with a
simple single-line regular expression.