Fix Android worker compilation of some engines & add FluidSynth
Android changed some system APIs like atof, rand, and srand in
various different versions of the Android API, and will not link
successfully if the wrong API version is used during compilation
since the included headers will have the wrong signatures. clang++
is a wrapper around the actual C++ compiler which sets them
correctly, so it is critical to always use this instead of the c++
binary. This caused linking to fail in several engines, at least
sword25 and wintermute.

Also, the directory where libraries are built currently is not
part of the default search path when using the clang++ wrapper, so
the CFLAGS/LDFLAGS needed to be corrected for that as well. This
should probably be changed in the future to build into the sysroot,
and for the moment I am just making this change to be able to move
forward with creating a release and finishing up with the Buildbot
work for now.

Finally, using FluidSynth Lite, which is FluidSynth ported to use
C11 instead of glib which gets rid of all of the difficult-to-build
FluidSynth dependencies like glib (which then requires libffi and
libiconv and libintl and etc.).